gl-util: Add a new crate that holds GL utilities
We wan't to be able to use the error framework across diferent
greenlight crates. The util crate is a good place to store some common
functionalities.
