+++ libs/gl-util/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "gl-util"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+bytes = "1.10"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1"
+tonic = "0.11"
